Html & jsx setup, TodoList component created
Start by adding the html elements to App.js and creating the <TodoList/> component then importing it into the App.js file. The html and TodoList component are all wrapped in a open and closed <div> element. Components are functions that return jsx. They also make up the visual layer of our application and allow the dev to split the application into different pieces