useState
React manages state inside of it's app. A state is a js object {name: 'value'} pair used to represent info about a specific component. Once class was the only format used to set state but now the dev can use hooks like useState to create component states. The todos is the current state and the setTodos is used to change the current state. useState must be imported to use it