props
function of a prop or property is to allow the parent component to pass data to the it's children. The prop being passed it 'Todo 1'&'Todo 2'. They are passed to the variable {todos} which = todoList which all all in the <TodoList/> component. In the component the values can be displayed, modified etc. Props are used in child components once  passed down.