prop keys, drilling values to child via prop
Todo component is the child of TodoList component which is child of App.js. The 3 prop values is sent to the todoList then TodoList component pass value from todoList to todo which the map function uses and displayed by the Todo component but there is a problem each prop value must be unique so 'key' is used so there won't be an error in the console.