state object
object, which is a {name:'value'} pair is used in the useState params. There are 3 name value pairs, id:1 which has a numeral value, name:'Todo 1' which has a string value, & complete:false which has a boolean value. id:1 is used by the map function in the TodoList component then the name variable is used in the Todo component