useEffect, useRe
The useEffect has two inputs to put in, the first is what you want to do which is inserted into the {} and the second one is when you want to do it which is inserted into the []. if the [] is empty the code only runs when the app has mount. The useEffect hook covers the lifecycle: mount, update, and unmount. useRef hook is used to reference the elements